% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_point_when_nonmissing.R
\name{add_point_when_nonmissing}
\alias{add_point_when_nonmissing}
\title{add_point_when_nonmissing}
\usage{
add_point_when_nonmissing(
  mod_summary,
  observed_data,
  group_col,
  year_col = year,
  time_counter = ndecade
)
}
\arguments{
\item{mod_summary}{A model summary created by \code{summarise_trends} or a related function.}

\item{observed_data}{The data used to fit the model summarised in mod_summary.}

\item{group_col}{The grouping variable to add points by.}

\item{year_col}{The column containing ISSP waves.}

\item{time_counter}{The column counting time.}
}
\description{
Adds a column containing values along the median regression line for each group when it is present in a survey wave and NA otherwise.
}
