% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_group_size.R
\name{calc_group_size}
\alias{calc_group_size}
\title{calc_group_size}
\usage{
calc_group_size(
  data,
  group,
  by = NULL,
  ids = NULL,
  probs = NULL,
  strata = NULL,
  fpc = NULL,
  weights = NULL,
  nest = FALSE
)
}
\arguments{
\item{data}{A data set object.}

\item{group}{The variable on which to calculate counts and proportions.}

\item{by}{Optional groups to nest observations by (e.g., attitude item, 
survey wave, country).}

\item{ids}{Variables specifying cluster ids from largest level to smallest 
level (leaving the argument empty, NULL, 1, or 0 indicate no clusters).}

\item{probs}{Variables specifying cluster sampling probabilities.}

\item{strata}{Variables specifying strata.}

\item{fpc}{Variables specifying a finite population correct 
(see \code{\link[survey]{svydesign}}).}

\item{weights}{Variables specifying weights (inverse of probability).}

\item{nest}{If \code{TRUE}, relabel cluster ids to enforce nesting within 
strata.}
}
\description{
Calculates the survey weighted count and proportion of observations on a 
variable.
}
