% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccode_to_cntry.R
\name{ccode_to_cntry}
\alias{ccode_to_cntry}
\title{ccode_to_cntry}
\usage{
ccode_to_cntry(x, custom_match = c(GB = "Great Britain"))
}
\arguments{
\item{x}{A vector of two-character ISO country codes.}

\item{custom_match}{A set of custom values mapping two-character ISO 
country codes to country names (see \code{\link[countrycode]{countrycode}}).}
}
\description{
Converts a vector of two-character ISO country codes to country names
}
