% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cntry_item_filter.R
\name{cntry_item_filter}
\alias{cntry_item_filter}
\title{cntry_item_filter}
\usage{
cntry_item_filter(
  data,
  att_col,
  cntry_col,
  wave_col = year,
  att_col_filter =
    "att_item \%in\% c(\\n      'grhseff2', 'cutenvir', 'prenvir', \\n      'taxenvir', 'privent', 'incdiff'\\n      )"
)
}
\arguments{
\item{data}{A data set object.}

\item{att_col}{A column with the names of attitude items.}

\item{cntry_col}{A column with the names of countries.}

\item{wave_col}{A column with the year of survey waves.}

\item{att_col_filter}{A string specifying a filter condition for the attitude items.}
}
\description{
Subsets to specified attitude items that are asked more than once in a country.
}
