% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparativeEnvPol-package.R
\docType{package}
\name{ComparativeEnvPol-package}
\alias{ComparativeEnvPol}
\alias{ComparativeEnvPol-package}
\title{ComparativeEnvPol: Functions and data for the analysis of political polarization on climate change using ISSP data}
\description{
This package is the repository of replication code and data for an ongoing project analyzing comparative trends in the political polarization of climate change using data from the International Social Survey Programme.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dcaldwellphd/ComparativeEnvPol}
  \item Report bugs at \url{https://github.com/dcaldwellphd/ComparativeEnvPol/issues}
}

}
\keyword{internal}
