% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ndecade.R
\name{count_ndecade}
\alias{count_ndecade}
\title{count_ndecade}
\usage{
count_ndecade(data, year_col, from = 1993)
}
\arguments{
\item{data}{A data set object.}

\item{year_col}{A column containing survey year.}

\item{from}{A numeric value to count time from.}
}
\description{
Adds a variable counting the number of decades from a specified year.
}
