% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efx_to_efx.R
\name{efx_to_efx}
\alias{efx_to_efx}
\title{efx_to_efx}
\usage{
efx_to_efx(
  baseline = mp_grhseff2_efx,
  baseline_obs = multiparty_pol,
  baseline_x = x,
  baseline_label = "baseline",
  comparisons,
  comparison_obs = multiparty_pol,
  comparison_x = x,
  comparison_labels = "comparison",
  add_stars_on = NULL
)
}
\arguments{
\item{baseline}{A data frame containing the summary of the posterior 
distribution of country effects from the baseline model.}

\item{baseline_obs}{A data frame containing the observed data used to fit 
the the baseline model.}

\item{baseline_x}{The variable used to model time in the baseline, passed
to the \code{time_counter} argument of 
\code{\link{add_point_when_nonmissing}}.}

\item{baseline_label}{A string to label rows summarizing the baseline model.}

\item{comparisons}{A single data frame or list of data frames containing 
the summary of the posterior distribution of country effects from the 
comparison models.}

\item{comparison_obs}{A data frame containing the observed data used to fit 
comparison models.}

\item{comparison_x}{The variable used to model time in the comparison, passed
to the \code{time_counter} argument of 
\code{\link{add_point_when_nonmissing}}.}

\item{comparison_labels}{A string or vector of strings to label rows 
summarizing the comparison model(s).}

\item{add_stars_on}{A variable indicating which country-slopes have a 
credible interval above zero (see details).}
}
\description{
Automates the data wrangling required to compare country effects from 
different models (see details).
}
\details{
Many plots in the paper compare climate concern polarization trends with 
something else, such as economic partisan polarization or the economic 
ideological constraint of climate-related attitudes. Joining summaries of 
the posterior distributions from different models involves a lot of 
repetitive data wrangling, which is automated by the \code{efx_to_efx} 
function.

The output is ordered by the median slope for countries from the baseline 
model, which is the climate danger partisan polarization model by default.
You can create a plot with a different order of countries by supplying a
different posterior summary to the \code{baseline} argument.

The \code{\link{summarise_trends}} function adds dummy variables to
the posterior summary of a model, indicating which groups have 90% and
95% credible intervals above zero. Providing one of these variables to
the \code{add_stars_on} argument of \code{efx_to_obs} will paste an
asterisk to country names showing that the credible interval of their
slope is positive in the baseline model.
}
