% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{make_plot}
\usage{
make_plot(
  data,
  x = x,
  y = mid,
  color = NULL,
  fill = pol_type,
  linetype = pol_type,
  shape = NULL,
  group = NULL,
  facet_vars = vars(cntry),
  ncol = NULL,
  points = NULL,
  line = mid,
  ci_upper = upper95,
  ci_lower = lower95,
  hline = 0,
  missingness_var = point_when_nonmissing,
  breaks = c("baseline", "comparison"),
  linetype_values = c(baseline = "solid", comparison = "dashed"),
  fill_values = c(baseline = "navyblue", comparison = "pink1"),
  group_values = NULL,
  color_values = NULL,
  shape_values = NULL,
  labels = NULL,
  ndecade_to_years = TRUE,
  ylab = "bquote('Adjusted ' ~ R^2)",
  ylim = c(-0.01, 0.24)
)
}
\arguments{
\item{data}{A data set containing the information to be plotted.}

\item{x}{Name of the column for the x-axis.}

\item{y}{Name of the column for the y-axis.}

\item{color}{Name of the column for an optional color aesthetic.}

\item{fill}{Name of the column for an optional fill aesthetic.}

\item{linetype}{Name of the column for an optional linetype aesthetic.}

\item{shape}{Name of the column for an optional shape aesthetic.}

\item{group}{Name of the column for an optional group aesthetic.}

\item{facet_vars}{A set of variables or expressions quoted by `ggplot2::vars` and defining faceting groups.}

\item{ncol}{An option to set the number of panels in a row when using facet_wrap.}

\item{points}{Name of the variable containing values of the points to draw for each group.}

\item{line}{Name of the variable containing values of the line to draw for each group.}

\item{ci_upper}{Name of the variable containing the upper bound of the credible interval of the median regression line for each group.}

\item{ci_lower}{Name of the variable containing the lower bound of the credible interval of the median regression line for each group.}

\item{hline}{A value along which to draw a horizontal line on the y-axis.}

\item{missingness_var}{A variable containing median values along the regression line for each group, for when you want to make the line and credible interval less promiment outside the range of available data.}

\item{breaks}{A character vector of breaks.}

\item{linetype_values}{A set of aesthetic values mapping linetypes to breaks.}

\item{fill_values}{A set of aesthetic values mapping fills to breaks.}

\item{group_values}{A set of aesthetic values mapping groups to breaks.}

\item{color_values}{A set of aesthetic values mapping colors to breaks.}

\item{shape_values}{A set of aesthetic values mapping shapes to breaks.}

\item{labels}{A character vector of labels for the breaks.}

\item{ndecade_to_years}{Logical. Do you want to display the number of decades as years on the x-axis?}

\item{ylab}{A string or expression to use as the y-axis label.}

\item{ylim}{A vector of values to use as the y-axis limits.}
}
\description{
A wrapper around \code{plot_structure}, \code{plot_layers}, and \code{plot_scales}, with default values set to common use cases in the paper.
}
