% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layers.R
\name{plot_layers}
\alias{plot_layers}
\title{plot_layers}
\usage{
plot_layers(
  p,
  points = NULL,
  line = NULL,
  ci_upper = NULL,
  ci_lower = NULL,
  hline = NULL,
  missingness_var = NULL
)
}
\arguments{
\item{p}{A \code{ggplot2} object.}

\item{points}{Name of the variable containing values of the points to draw 
for each group.}

\item{line}{Name of the variable containing values of the line to draw 
for each group.}

\item{ci_upper}{Name of the variable containing the upper bound of the 
credible interval of the median regression line for each group.}

\item{ci_lower}{Name of the variable containing the lower bound of the 
credible interval of the median regression line for each group.}

\item{hline}{A value along which to draw a horizontal line on the y-axis.}

\item{missingness_var}{A variable containing median values along the 
regression line for each group, for when you want to make the line and 
credible interval less promiment outside the range of available data.}
}
\description{
Takes a \code{ggplot2} object as input and adds specified geom 
layers that can highlight missingness in the data.
}
