% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scales.R
\name{plot_scales}
\alias{plot_scales}
\title{plot_scales}
\usage{
plot_scales(
  p,
  breaks = NULL,
  linetype_values = NULL,
  fill_values = NULL,
  group_values = NULL,
  color_values = NULL,
  shape_values = NULL,
  labels = NULL,
  ndecade_to_years = FALSE,
  ylab = NULL,
  ylim = NULL
)
}
\arguments{
\item{p}{A ggplot2 object.}

\item{breaks}{A character vector of breaks.}

\item{linetype_values}{A set of aesthetic values mapping linetypes to breaks.}

\item{fill_values}{A set of aesthetic values mapping fills to breaks.}

\item{group_values}{A set of aesthetic values mapping groups to breaks.}

\item{color_values}{A set of aesthetic values mapping colors to breaks.}

\item{shape_values}{A set of aesthetic values mapping shapes to breaks.}

\item{labels}{A character vector of labels for the breaks.}

\item{ndecade_to_years}{Logical. Do you want to display the number of decades as years on the x-axis?}

\item{ylab}{A string or expression to use as the y-axis label.}

\item{ylim}{A vector of values to use as the y-axis limits.}
}
\description{
Takes a ggplot2 object as input and sets scales and labels for axes and geom layers
}
