% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_structure.R
\name{plot_structure}
\alias{plot_structure}
\title{plot_structure}
\usage{
plot_structure(
  data,
  x,
  y,
  color = NULL,
  fill = NULL,
  linetype = NULL,
  shape = NULL,
  group = NULL,
  facet_vars = NULL,
  ncol = NULL
)
}
\arguments{
\item{data}{A data set containing the information to be plotted.}

\item{x}{Name of the column for the x-axis.}

\item{y}{Name of the column for the y-axis.}

\item{color}{Name of the column for an optional color aesthetic.}

\item{fill}{Name of the column for an optional fill aesthetic.}

\item{linetype}{Name of the column for an optional linetype aesthetic.}

\item{shape}{Name of the column for an optional shape aesthetic.}

\item{group}{Name of the column for an optional group aesthetic.}

\item{facet_vars}{A set of variables or expressions quoted by 
\code{ggplot2::vars} and defining faceting groups.}

\item{ncol}{An option to set the number of panels in a row when using 
\code{facet_wrap}.}
}
\description{
Creates a \code{ggplot2} object with any specified faceting and aesthetics.
}
