% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_wave.R
\name{prepare_wave}
\alias{prepare_wave}
\title{prepare_wave}
\usage{
prepare_wave(
  data,
  cntry_col,
  year_value,
  weight_col,
  rm_cntry = "Northern Ireland",
  custom_ccode_to_cntry = c(`Germany-East - D-E` = "DE", `Germany-West - D-W` = "DE",
    `D-E-Germany-East` = "DE", `D-W-Germany-West` = "DE"),
  custom_cntry_to_ccode = c(GB = "Great Britain")
)
}
\arguments{
\item{data}{A wave of ISSP data.}

\item{cntry_col}{The column with country names.}

\item{year_value}{A numeric value to add as \code{year} to the output.}

\item{weight_col}{The column with post-stratification weights.}

\item{rm_cntry}{Countries to remove from the wave (see details).}

\item{custom_ccode_to_cntry}{A set of custom values mapping country names
to country codes (see details).}

\item{custom_cntry_to_ccode}{A set of custom values mapping country codes
to country names (see details).}
}
\description{
Performs a number of miscellaneous operations on a wave of ISSP data,
establishing consistent country codes and names, creating id and year
columns for keeping track of observations when joining waves together, and
applying a consistent name and numeric class to the column containing
post-stratification weights.
}
\details{
To remove countries from a wave of ISSP data, provide \code{rm_cntry} with
a pattern to match based on country names used by the ISSP in that wave.
For example, default searches for "Northern Ireland" and removes
corresponding rows. Remove multiple countries using a pattern like
"Nothern Ireland|...".

The function uses \code{countrycode} to assign consistent two-character ISO
codes based on country names used by the ISSP. However, some ISSP values
require a custom mapping, such as the labels for East and West Germany
before reunification.

Similarly, the function uses \code{countrycode} on the newly created ISO
codes to return country names that may be more consistent than values used
by the ISSP across waves. The \code{custom_cntry_to_ccode} makes it
possible to specify a custom mapping for these country names. For example,
the default hard codes "GB" to "Great Britain" (rather than "United
Kingdom"), reflecting the fact that Northern Ireland is excluded by default.
}
\seealso{
\code{\link[countrycode]{countrycode}}
}
