% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ideology.R
\name{process_ideology}
\alias{process_ideology}
\title{process_ideology}
\usage{
process_ideology(data, issp_name, csv = issp_env_vars)
}
\arguments{
\item{data}{A wave of ISSP data.}

\item{issp_name}{Name of the column in the CSV file containing ISSP 
attitude item names for the wave.}

\item{csv}{The csv file used to recode the wave of ISSP data.}
}
\description{
Recodes attitude items in an ISSP wave using values from a CSV file.
}
