% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_alt_usa_pid.R
\name{recode_alt_usa_pid}
\alias{recode_alt_usa_pid}
\title{recode_alt_usa_pid}
\usage{
recode_alt_usa_pid(data, usa_pid)
}
\arguments{
\item{data}{A wave of ISSP data.}

\item{usa_pid}{A column containing responses to the
American party identification strength question.}
}
\description{
Provides two alternative recodings of the American party 
identification strength variable.
}
