% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_item_type_trends.R
\name{summarise_item_type_trends}
\alias{summarise_item_type_trends}
\title{summarise_item_type_trends}
\usage{
summarise_item_type_trends(
  mod,
  x = "ndecade",
  x_vals = c(0, 0.7, 1.7, 2.7),
  z1 = "is_env_cogaffTRUE",
  z2 = "is_climate_cogaffTRUE",
  cntry_var = "ccode"
)
}
\arguments{
\item{mod}{A stanreg object.}

\item{x}{A string containing the name of the variable counting time.}

\item{x_vals}{A vector containing values of x at which to calculate the median and credible interval.}

\item{z1}{A string containing the name of the first dummy variable.}

\item{z2}{A string containing the name of the second dummy variable.}

\item{cntry_var}{A string containing the name of the variable used to model country effects.}
}
\description{
This is a special case of \code{\link{summarise_trends}} for summarising trends separated by item type (see details).
}
\details{
This function is specifically meant for a three-way grouping of attitude items based on topic-expression (environment-conative, environment-cognitive/affective, climate-cognitive/affective).
}
