% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_trends.R
\name{summarise_trends}
\alias{summarise_trends}
\title{summarise_trends}
\usage{
summarise_trends(
  mod,
  x = "ndecade",
  x_vals = c(0, 0.7, 1.7, 2.7),
  control_var = NULL,
  control_val = NULL
)
}
\arguments{
\item{mod}{A stanreg object.}

\item{x}{A string containing the name of the variable counting time.}

\item{x_vals}{A vector containing values of x at which to calculate the 
median and credible interval.}

\item{control_var}{A string containing the name of a control variable column.}

\item{control_val}{A numeric value for fitting group-level trends 
(e.g., the overall mean in the data).}
}
\description{
Takes a multilevel \code{stanreg} object as input and summarises the
posterior distribution of linear time trends for each group in the
model (see details).
}
\details{
The main approach to modelling polarization in the paper is to predict
some statistical outcome using a multilevel model with the number of
decades from 1993 as a scalar independent variable. This function
summarises the rergession line of each group along values of this 
independent variable. For models where polarization is predicted 
using dummy variables for each year, see \code{\link{summarise_years}}.
}
