% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_years.R
\name{summarise_years}
\alias{summarise_years}
\title{summarise_years}
\usage{
summarise_years(mod, ci_width = 0.95)
}
\arguments{
\item{mod}{A stanreg object.}

\item{ci_width}{The width of the credible interval to be reported.}
}
\description{
Takes a multilevel \code{stanreg} object as input and summarises the
posterior distribution of year effects for each group in the
model (see details).
}
\details{
The supplementary information for the paper considers nonlinear
changes in partisan polarization using a multilevel model with
dummy variables for each year. This function summarises the change
in partisan polarization for each group in 2000, 2010, and 2020
relative to 1993. For models where polarization is predicted using
a scalar variable counting time in decades, see 
\code{\link{summarise_decades}}.
}
