% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccode_to_cntry.R
\name{ccode_to_cntry}
\alias{ccode_to_cntry}
\title{ccode_to_cntry}
\usage{
ccode_to_cntry(
  data,
  input_col,
  output_col = cntry,
  custom_match = c(GB = "Great Britain")
)
}
\arguments{
\item{data}{A data set object.}

\item{input_col}{A column containing two-character ISO country codes.}

\item{output_col}{The name to give the resulting column of country names.}

\item{custom_match}{A set of custom values mapping two-character ISO country codes to country names (see \code{\link[countrycode]{countrycode}}).}
}
\description{
Converts a column of two-character ISO country codes to country names
}
