% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efx_to_efx.R
\name{efx_to_efx}
\alias{efx_to_efx}
\title{efx_to_efx}
\usage{
efx_to_efx(
  baseline = mp_grhseff2_efx,
  baseline_obs = multiparty_pol,
  baseline_label = "baseline",
  comparisons,
  comparison_obs = multiparty_pol,
  comparison_labels = "comparison"
)
}
\arguments{
\item{baseline}{A data frame containing the summary of the posterior distribution of country effects from the baseline model.}

\item{baseline_obs}{A data frame containing the observed data used to fit the the baseline model.}

\item{baseline_label}{A string to label rows summarizing the baseline model.}

\item{comparisons}{A single data frame or list of data frames containing the summary of the posterior distribution of country effects from the comparison models.}

\item{comparison_obs}{A data frame containing the observed data used to fit comparison models.}

\item{comparison_labels}{A string or vector of strings to label rows summarizing the comparison model(s).}
}
\description{
Automates the data wrangling required to compare country effects from different models (see details).
}
\details{
Many plots in the paper compare climate concern polarization trends with something else, such as economic partisan polarization or the economic ideological constraint of climate-related attitudes. Joining summaries of the posterior distributions from different models involves a lot of repetitive data wrangling, which is automated by the \code{efx_to_efx} function.
}
