% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efx_to_obs.R
\name{efx_to_obs}
\alias{efx_to_obs}
\title{efx_to_obs}
\usage{
efx_to_obs(
  obs_data = multiparty_pol,
  efx_data = mp_grhseff2_efx,
  filter_obs_to = "grhseff2"
)
}
\arguments{
\item{obs_data}{A data frame containing the observed data used to fit 
the model.}

\item{filter_obs_to}{An optional string to filter the observed data to a 
selection of attitude items.}

\item{baseline}{A data frame containing the summary of the posterior 
distribution of country effects from the model.}
}
\description{
Automates the data wrangling required to compare country effects from a 
model with the observed data used to fit it.
}
