% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv.R
\name{load_csv}
\alias{load_csv}
\title{load_csv}
\usage{
load_csv(csv)
}
\arguments{
\item{csv}{Name of the CSV file}
}
\description{
ComparativeEnvPol uses CSV files to recode attitude and party support items, 
which are bundled in its `inst/extdata` directory. 
This function makes them easy to access
}
