% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_trends.R
\name{summarise_trends}
\alias{summarise_trends}
\title{summarise_trends}
\usage{
summarise_trends(
  mod,
  x = "ndecade",
  x_vals = c(0, 0.7, 1.7, 2.7),
  control_var = NULL,
  control_val = NULL
)
}
\arguments{
\item{mod}{A stanreg object.}

\item{x}{A string containing the name of the variable counting time.}

\item{x_vals}{A vector containing values of x at which to calculate the median and credible interval.}

\item{control_var}{A string containing the name of a control variable column.}

\item{control_val}{A numeric value for fitting group-level trends (e.g., the overall mean in the data).}
}
\description{
Takes a multilevel \code{stanreg} object as input and returns the median and credible interval of time trends for each group.
}
